Project cleanup and documentation work
Old files have been removed that are no longer used in the program,
these are either obsolete or were only included for initial testing /
development.

The UML has had a large update and is almost up to date with the current
progress of the project. Some final updates are still required.