Working state, moving camera and player moves
The camera issue is now fixed, this was due to forgetting to include
inView in the shader + in the meshRenderer.

The player movement has been resolved, check main.cpp for example of how
this works, now needs to be linked with user input.

Keyboard input fixed, was increasing the wrong iterator on an internal
loop throwing it out of the loops range. changed inner loop to iterate
the correct variable now.